refactored test package
