+++ src/test/Mustafa.java
@@ -1,2 +1,4 @@
+package test;
+
 public class Mustafa {
 }
